requirements.txt
# ===========================
# Django Core Dependencies
# ===========================
Django>=4.2,<5.0  # Django framework (version 4.2 up to 5.0)

# ===========================
# Django Extensions & Add-ons
# ===========================
django-filter==24.3  # Filtering support for Django REST framework
django-mail-templated  # Email templating for Django

# ===========================
# Django REST Framework (DRF)
# ===========================
djangorestframework==3.15.2  # Django REST Framework for building APIs
djangorestframework-simplejwt  # JWT authentication for DRF

# ===========================
# API Documentation & Validation
# ===========================
drf-yasg[validation]  # Swagger & Redoc documentation for DRF

# ===========================
# Configuration & Environment Management
# ===========================
python-decouple==3.8  # Managing environment variables & settings


#==============================
#Reformatting
#==============================
balck
flake8

#==============================
#tests
#==============================
pytest
pytest-django



